+++ src/main/java/lazyae2/network/ModNetwork.java
@@ -27,5 +27,6 @@ public final class ModNetwork {
         CHANNEL.registerMessage(PacketTerminalRow.Handler.class, PacketTerminalRow.class, 3, Side.SERVER);
         CHANNEL.registerMessage(PacketTerminalAmount.Handler.class, PacketTerminalAmount.class, 4, Side.SERVER);
         CHANNEL.registerMessage(PacketTerminalFilter.Handler.class, PacketTerminalFilter.class, 5, Side.SERVER);
+        CHANNEL.registerMessage(PacketAssemblerPatterns.Handler.class, PacketAssemblerPatterns.class, 6, Side.CLIENT);
     }
 }